change calc of statuseffect duration

+++ CSharpSourceCode/BattleMechanics/StatusEffect/StatusEffect.cs
@@ -5,7 +5,7 @@ namespace TOR_Core.BattleMechanics.StatusEffect
     public class StatusEffect
     {
         public Agent ApplierAgent = null;
-        public int CurrentDuration = 0;
+        public float CurrentDuration = 0;
         public StatusEffectTemplate Template { get; }
 
         public StatusEffect(StatusEffectTemplate template)
